RAGFLOW+LANGCHAIN/requirements_gui.txt
# RAGFlow GUI 依赖包
# 核心依赖
tkinter
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# 图像处理 (用于图标和界面美化)
Pillow>=9.0.0

# 网络请求
requests>=2.25.0

# 环境变量管理
python-dotenv>=0.19.0

# 数据处理
numpy>=1.21.0

# 向量数据库支持
faiss-cpu>=1.7.0
chromadb>=0.4.0

# 异步支持
asyncio